Make checkguard easier to use
Some projects have very inconsistent include guard rules, so
perhaps a better default would be to just check that there is some
guard-looking thing in each file. We can add an option to supply a
pattern to check against to restore the old behaviour (which I found
caught mistakes in projects with consistent rules).