Reject defines that are not guards
Macro definitions can be confused with include guards when we're
guessing the symbol. This is more likely to be a problem now that we
accept include guards with values (#15), but could happen regardless.

If the define is followed immediately by an #endif, there is no content
being guarded and it's probably just a define.