Remove version bump disclaimer
This disclaimer was mostly to soothe my nerves when releasing v2.0. I
might as well bump the version number if I make major breaking changes.

The big remaining design problem is that without an escape character,
the special character handling is incomplete. For example, it's not
possible to specify an argument containing both ' and " characters.