Add support for negative values to path
This is really handy because otherwise you pretty much have to change
directory into just the right spot to use -p with path effectively.

I frequently found myself wanting to just run guardonce over some
project and needing to strip the first N directories from the path
because they were merely part of the relative path to the project.
This does exactly that.

It's a tad weird, but the feature is logically consistent if you think
of this as specifying the number of parent directories to keep in the
same style as python indexing. More importantly, though, the feature is
just very useful.