Handle UTF-8 with BOM
Strip the first few bytes of the file if it matches the BOM character.
This should make UTF-8 with BOM files work as expected, and it seems
unlikely to cause problems. While it's possible the header *could* be
a CP-1252 encoded file starting with 'ï»¿'... It seems unlikely that
program would compile. In fact, any non-ASCII sequence of characters
at the start of the header probably wouldn't compile.

Fixes #23.