Use the exact same test data on every platform
Settings like core.autocrlf can cause platform-specific changes to the
test data if .gitattributes are left unspecified. That's a big problem
for tests that are trying to check how line endings are handled.

guardonce should work exactly the same on Windows as on Linux or OSX,
so it should have exactly the same tests.